graph: move tool conversion tests
